+++ spec/controllers/v2/todos_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe V2::TodosController, type: :controller do
+
+end
