+++ spec/factories/todos.rb
@@ -1,4 +1,4 @@
-FactoryBot.define do
+  FactoryBot.define do
   factory :todo do
     title { Faker::Lorem.word }
     created_by { Faker::Number.number(digits: 10) }
